backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-RESTful==0.3.10
PyMySQL==1.1.0
cryptography==41.0.7
python-dotenv==1.0.0
requests==2.31.0
reportlab==4.0.7
PyPDF2==3.0.1
ollama==0.1.7
marshmallow==3.20.1
flask-marshmallow==0.15.0
marshmallow-sqlalchemy==0.29.0
Werkzeug==3.0.1
gunicorn==21.2.0